+++ src/main.rs
@@ -1,4 +1,7 @@
 mod app;
+mod chat_log;
+mod default_keybinds;
+mod event_bus;
 mod socket_client;
 mod tui_framework;
 mod ui;
@@ -16,30 +19,16 @@ use ratatui::prelude::{CrosstermBackend, Terminal};
 use std::io::stdout;
 
 use crate::app::App;
-use crate::update::update;
+use crate::event_bus::dispatch;
 use crate::user_config::load_config;
 use tui_framework::*;
 
 async fn setup() -> Result<(App, Tui)> {
     let terminal = Terminal::new(CrosstermBackend::new(stdout()))?;
-    let mut tui = Tui::from_conf(
-        terminal,
-        TuiConf {
-            update_freq: 30.0,
-            ..TuiConf::default()
-        },
-    )
-    .default_client();
+    let mut tui = Tui::from_conf(terminal, TuiConf::default()).default_client();
 
     let mut app = App::new(load_config().await);
-    let (client, token) = match tui
-        .connect(ClientMsg {
-            token: None,
-            body: ClientMsgBody::Login(app.username.clone()),
-            timestamp: Timestamp::from(Utc::now()),
-        })
-        .await
-    {
+    let (client, token) = match tui.connect(login_msg(&app)).await {
         Some(x) => x,
         None => panic!("Could not retrieve token from server"),
     };
@@ -54,47 +43,20 @@ async fn setup() -> Result<(App, Tui)> {
     Ok((app, tui))
 }
 
+fn login_msg(app: &App) -> ClientMsg {
+    ClientMsg {
+        token: None,
+        body: ClientMsgBody::Login(app.username.clone()),
+        timestamp: Timestamp::from(Utc::now()),
+    }
+}
+
 async fn run() -> Result<()> {
     let (mut app, mut tui) = setup().await?;
 
     while !app.should_quit {
         let event = tui.next().await?;
-        if let Event::Render = event {
-            tui.draw(&mut app)?;
-        }
-        update(&mut app, event.clone());
-
-        match event {
-            Event::Send {
-                token,
-                timestamp,
-                contents,
-                ..
-            } => {
-                let msg = ClientMsg {
-                    token: Some(token),
-                    body: ClientMsgBody::SendToRoom { contents },
-                    timestamp: Timestamp::from(timestamp),
-                };
-                tui.push_binary_msg_to_server(msg);
-            }
-            Event::ServerCommand {
-                token,
-                timestamp,
-                message_body,
-                ..
-            } => {
-                let server_msg = ClientMsg {
-                    token: Some(token),
-                    timestamp: Timestamp::from(timestamp),
-                    body: message_body,
-                };
-                tui.push_binary_msg_to_server(server_msg);
-            }
-            _ => {
-                log::info!("No handling for {event:?}");
-            }
-        }
+        dispatch(&mut app, &mut tui, event)?;
     }
 
     tui.exit()?;
@@ -105,8 +67,6 @@ async fn run() -> Result<()> {
 #[tokio::main]
 async fn main() -> Result<()> {
     _ = log2::open("log.txt").module(true).start();
-
-    log::error!("SANITY CHECK");
 
     let result = run().await;
