+++ src/update.rs
@@ -1,9 +1,13 @@
-use crate::app::Log;
 use crate::app::{App, Mode};
+use crate::chat_log::Log;
 use crate::tui_framework::Event;
 use chrono::{DateTime, Utc};
 use crossterm::event::KeyEvent;
-use marain_api::prelude::{ChatMsg, ServerMsg, ServerMsgBody, Status};
+use marain_api::prelude::{ChatMsg, ServerMsg, ServerMsgBody, Status, Timestamp};
+
+fn translate_ts(ts: Timestamp) -> DateTime<Utc> {
+    Into::<Option<DateTime<Utc>>>::into(ts).unwrap_or(Utc::now())
+}
 
 pub fn update(app: &mut App, event: Event) {
     match event {
@@ -30,11 +34,10 @@ pub fn update(app: &mut App, event: Event) {
             match bincode::deserialize::<ServerMsg>(&msg[..]) {
                 Ok(deserialized) => {
                     app.push_debug_log(deserialized.clone());
-                    let timestamp_dt = Into::<Option<DateTime<Utc>>>::into(deserialized.timestamp);
 
                     // Handle any errors
                     match deserialized.status {
-                        Status::Yes => {}
+                        Status::Yes => handle_server_msg(app, deserialized),
                         Status::No(error_msg) => {
                             app.push_log(Log::new("SERVER".into(), error_msg.clone()));
                             log::error!("The computer said no: {error_msg}");
@@ -43,47 +46,6 @@ pub fn update(app: &mut App, event: Event) {
                         Status::JustNo => {
                             app.push_log(Log::new("CLIENT".into(), "Failed to login".into()));
                             return;
-                        }
-                    }
-
-                    // These are all success responses from the server
-                    match deserialized.body {
-                        ServerMsgBody::LoginSuccess { token } => app.store_token(token),
-                        ServerMsgBody::ChatRecv {
-                            chat_msg:
-                                ChatMsg {
-                                    sender, content, ..
-                                },
-                            ..
-                        } => {
-                            app.push_log(
-                                Log::new(sender, content).at(timestamp_dt.unwrap_or(Utc::now())),
-                            );
-                        }
-                        ServerMsgBody::Empty => {
-                            let server_time = timestamp_dt.unwrap_or_else(|| {
-                                log::error!("Server did not supply time");
-                                Utc::now()
-                            });
-                            app.push_log(Log::new(
-                                "SERVER".into(),
-                                "The time is: ".to_string()
-                                    + &server_time.format("%Y-%m-%D %H:%M:%S").to_string(),
-                            ))
-                        }
-                        ServerMsgBody::RoomData { logs, .. } => {
-                            let chat_logs: Vec<Log> = logs
-                                .iter()
-                                .map(|cm| {
-                                    Log::new(cm.sender.clone(), cm.content.clone()).at(Into::<
-                                        Option<DateTime<Utc>>,
-                                    >::into(
-                                        cm.timestamp.clone(),
-                                    )
-                                    .unwrap_or(Utc::now()))
-                                })
-                                .collect();
-                            app.replace_logs(chat_logs);
                         }
                     }
                 }
@@ -96,5 +58,35 @@ pub fn update(app: &mut App, event: Event) {
             }
         }
         _ => {}
+    }
+}
+
+fn handle_server_msg(app: &mut App, deserialized: ServerMsg) {
+    let dt = translate_ts(deserialized.timestamp.clone());
+    // These are all success responses from the server
+    match deserialized.body {
+        ServerMsgBody::LoginSuccess { token } => app.store_token(token),
+        ServerMsgBody::ChatRecv {
+            chat_msg: ChatMsg {
+                sender, content, ..
+            },
+            ..
+        } => {
+            app.push_log(Log::new(sender, content).at(dt));
+        }
+        ServerMsgBody::Empty => app.push_log(Log::new(
+            "SERVER".into(),
+            "The time is: ".to_string() + &dt.format("%Y-%m-%D %H:%M:%S").to_string(),
+        )),
+        ServerMsgBody::RoomData { logs, .. } => {
+            let chat_logs: Vec<Log> = logs
+                .iter()
+                .map(|cm| {
+                    Log::new(cm.sender.clone(), cm.content.clone())
+                        .at(translate_ts(cm.timestamp.clone()))
+                })
+                .collect();
+            app.replace_logs(chat_logs);
+        }
     }
 }
